For multi-status response, promote the first erroring operations error code that caused the batch to fail

+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/Batch/BatchAsyncContainerExecutorTests.cs
@@ -32,7 +32,10 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestCleanup]
         public async Task CleanupAsync()
         {
-            await this.cosmosContainer.Database.DeleteAsync();
+            if (this.cosmosContainer != null)
+            {
+                await this.cosmosContainer.Database.DeleteAsync();
+            }
         }
 
         [TestMethod]
